Naive Start End Room Selection
Created a way to select a start and end room. Start room is selected via min.x, min.z condition while the end room is selected with the max.z, max.x.

An extension to make the selection more appropriate, a graph starting from the selected starting point will be created to find the longest path within the dungeon to place the end room.